requirements_waveshare.txt
# NFC Display System Requirements - Waveshare Version

# For Waveshare PN532 NFC HAT (UART)
pyserial==3.5
RPi.GPIO==0.7.1

# Web Interface Dependencies
Flask==2.3.2
Flask-SocketIO==5.3.4
python-socketio==5.9.0

# System dependencies (install via apt):
# sudo apt install python3-dev python3-pip
